refactor : BaseEntity 이동, Meeting 엔티티 명을 Group으로 변경
